refactor(assets/scripts): rename extension mjs to js
