+++ src/assets/scripts/gaOptout.js
