+++ src/assets/scripts/gaOptoutKeyStorage.js
